+++ app/src/main/java/ru/practicum/android/diploma/App.kt
@@ -3,13 +3,9 @@ package ru.practicum.android.diploma
 import android.app.Application
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.context.startKoin
-import ru.practicum.android.diploma.di.favoritesModule
-import ru.practicum.android.diploma.di.favoritesViewModelModule
-import ru.practicum.android.diploma.di.networkClientModule
+import ru.practicum.android.diploma.di.dataModule
+import ru.practicum.android.diploma.di.domainModule
 import ru.practicum.android.diploma.di.viewModelModule
-import ru.practicum.android.diploma.search.di.searchRepositoryModule
-import ru.practicum.android.diploma.share.di.shareModule
-import ru.practicum.android.diploma.di.vacancyDetailsRepositoryModule
 
 class App : Application() {
 
@@ -20,12 +16,8 @@ class App : Application() {
             androidContext(this@App)
             modules(
                 viewModelModule,
-                favoritesModule,
-                favoritesViewModelModule,
-                networkClientModule,
-                searchRepositoryModule,
-                vacancyDetailsRepositoryModule,
-                shareModule
+                domainModule,
+                dataModule,
             )
         }
     }
