--- app/src/main/java/ru/practicum/android/diploma/di/FavoritesViewModelModule.kt
@@ -1,12 +1,0 @@
-package ru.practicum.android.diploma.di
-
-import org.koin.androidx.viewmodel.dsl.viewModel
-import org.koin.dsl.module
-import ru.practicum.android.diploma.favorites.presentation.FavoritesViewModel
-
-val favoritesViewModelModule = module {
-
-    viewModel {
-        FavoritesViewModel(get())
-    }
-}
