--- app/src/main/java/ru/practicum/android/diploma/di/NetworkClientModule.kt
@@ -1,35 +1,0 @@
-package ru.practicum.android.diploma.di
-
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import org.koin.dsl.module
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import ru.practicum.android.diploma.search.data.api.HHApiService
-import ru.practicum.android.diploma.search.data.api.NetworkClient
-import ru.practicum.android.diploma.search.data.network.RetrofitNetworkClient
-
-const val BASE_URL = "https://api.hh.ru/"
-
-val networkClientModule = module {
-
-    single<HHApiService> {
-        val client = OkHttpClient.Builder().addInterceptor(
-            HttpLoggingInterceptor().apply {
-                setLevel(HttpLoggingInterceptor.Level.BODY)
-            }
-        ).build()
-
-        val retrofit = Retrofit.Builder()
-            .baseUrl(BASE_URL)
-            .client(client)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-
-        retrofit.create(HHApiService::class.java)
-    }
-
-    single<NetworkClient> {
-        RetrofitNetworkClient(get())
-    }
-}
