+++ app/src/main/java/ru/practicum/android/diploma/di/ViewModelModule.kt
@@ -2,10 +2,15 @@ package ru.practicum.android.diploma.di
 
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
+import ru.practicum.android.diploma.favorites.presentation.FavoritesViewModel
 import ru.practicum.android.diploma.search.presentation.SearchVacanciesViewModel
 import ru.practicum.android.diploma.vacancy.presentation.VacancyDetailsViewModel
 
 val viewModelModule = module {
+
+    viewModel {
+        FavoritesViewModel(get())
+    }
 
     viewModel {
         SearchVacanciesViewModel(get())
